backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sqlalchemy==2.0.22
asyncmy==0.2.9
pandas==1.5.3
spacy
requests
numpy==1.26.4
openai==0.27.0
python-dotenv==1.0.0
pydantic==1.8.2
sentence-transformers==2.1.0
dash==2.10.2
plotly==5.14.1
requests==2.32.3
scikit-learn==1.5.1
transformers==4.29.2
nltk==3.8.1
gensim==4.3.1

